chore: divide test files
